make JsFuncHandler an AlgebraM
Move code from MongoDbPlanner to JsFuncHandler.
Also simplfy implementation of mapFuncDerived because ExpandMapFunc.expand can now be called
